add jurisdiction disambiguators
