+++ db/migrate/20250626162435_add_disambiguator_to_jurisdictions.rb
@@ -1,0 +1,5 @@
+class AddDisambiguatorToJurisdictions < ActiveRecord::Migration[7.1]
+  def change
+    add_column :jurisdictions, :disambiguator, :string
+  end
+end
