+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2025_06_24_163718) do
+ActiveRecord::Schema[7.1].define(version: 2025_06_26_162435) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
   enable_extension "plpgsql"
@@ -224,6 +224,7 @@ ActiveRecord::Schema[7.1].define(version: 2025_06_24_163718) do
     t.string "external_api_state", default: "g_off", null: false
     t.boolean "inbox_enabled", default: false, null: false
     t.integer "heating_degree_days"
+    t.string "disambiguator"
     t.index ["prefix"], name: "index_jurisdictions_on_prefix", unique: true
     t.index ["regional_district_id"], name: "index_jurisdictions_on_regional_district_id"
     t.index ["slug"], name: "index_jurisdictions_on_slug", unique: true
